Release v0.7.11: fix the crashes, the lost settings and the frozen terminal
A pass over the whole program for bugs, and the fixes for what it turned up.

Three of them could end the process: a release manifest or a MUD list with a
field written as null (a JSON null overrides a non-nullable initializer, and
both were dereferenced in a background timer nothing was catching), and a JAWS
restart landing between one thread's check on the bound object and another
thread's call to it.

One could lose every setting a person had: a single hand-edited number outside
the range a terminal can be was answered by throwing the whole file away and
starting from the defaults, permanently, because the next save wrote over it.

One silently reverted a choice: the settings dialog edits a copy, and the window
copied every field back except the theme, so choosing dark lasted until the next
thing that saved anything.

The flood detection added in v0.7.9 never engaged for a program printing slower
than Windows' own 15.6 ms timer tick -- which is what the two failing flood tests
were saying. A build printing sixty lines a second was still spoken a batch at a
time, queueing behind itself exactly as it did before that feature existed.

A MUD could freeze the terminal for twenty seconds by naming a sound address
that never answers, because the fetch happened on the window's thread with the
sound board's lock held.

Also fixed: an article is no longer stripped off the middle of a game name, so
"Thera" and "Ra" stay two different games rather than sharing each other's
player counts; a trigger waiting out its own cooldown no longer stops the
triggers listed below it; a submitted line and its Return can no longer be
interleaved by another submission; a telnet subnegotiation that is never closed
no longer grows without bound and swallows the rest of the connection; a sound
named "NUL.wav" or "COM1.wav" no longer reaches the device Windows answers with;
a mistyped shell name no longer takes the process down and leaks the pseudo
console; and a copy taken while another program holds the clipboard open no
longer ends the terminal.

942 tests and 33 replays pass, including 27 new ones -- one for each of the
smaller bugs.

+++ src/BlindTerm.Core/Mud/MudSorting.cs
@@ -82,14 +82,17 @@ public static class MudSorting
         perPage = Math.Max(1, perPage);
         page = Math.Max(1, page);
 
-        int start = (page - 1) * perPage;
+        // Worked out in long. page is only bounded below, and (int.MaxValue - 1) * perPage
+        // overflows to a negative start -- which reads as "before the first page" and quietly
+        // returns the first one instead of nothing.
+        long start = (long)(page - 1) * perPage;
         if (start >= games.Count)
             return new MudDirectoryPage { Games = [], Page = page, PerPage = perPage, Total = games.Count };
 
-        int length = Math.Min(perPage, games.Count - start);
+        int length = (int)Math.Min(perPage, games.Count - start);
         return new MudDirectoryPage
         {
-            Games = [.. games.Skip(start).Take(length)],
+            Games = [.. games.Skip((int)start).Take(length)],
             Page = page,
             PerPage = perPage,
             Total = games.Count,
